CLN: Move common dataclasses to ecl_grid
