+++ src/xtgeo/grid3d/_grdecl_grid.py
@@ -20,272 +20,22 @@ And ignore ECHO and NOECHO keywords. see _grid_format for the details
 of how these keywords are layed out in a text file, and see GrdeclGrid
 for how the grid geometry is interpreted from these keywords.
 """
-from dataclasses import InitVar, astuple, dataclass, fields
-from enum import Enum, auto, unique
+from dataclasses import InitVar, dataclass
 from typing import Optional, Tuple
 
 import numpy as np
 from ecl_data_io import Format, lazy_read, write
 
-from ._ecl_grid import EclGrid
-from ._grdecl_format import match_keyword, open_grdecl
-
-
-@unique
-class GridRelative(Enum):
-    """GridRelative is the second value given GRIDUNIT keyword.
-
-    MAP means map relative units, while
-    leaving it blank means relative to the origin given by the
-    MAPAXES keyword.
-    """
-
-    MAP = auto()
-    ORIGIN = auto()
-
-    def to_grdecl(self):
-        if self == GridRelative.MAP:
-            return "MAP"
-        else:
-            return ""
-
-    def to_bgrdecl(self):
-        return self.to_grdecl().ljust(8)
-
-    @classmethod
-    def from_grdecl(cls, unit_string):
-        if match_keyword(unit_string, "MAP"):
-            return cls.MAP
-        else:
-            return cls.ORIGIN
-
-    @classmethod
-    def from_bgrdecl(cls, unit_string):
-        if isinstance(unit_string, bytes):
-            return cls.from_grdecl(unit_string.decode("ascii"))
-        return cls.from_grdecl(unit_string)
-
-
-@unique
-class Order(Enum):
-    """Either increasing or decreasing.
-
-    Used for the grdecl keywords INC and DEC
-    respectively.
-    """
-
-    INCREASING = auto()
-    DECREASING = auto()
-
-    def to_grdecl(self):
-        return str(self.name)[0:3]
-
-    def to_bgrdecl(self):
-        return self.to_grdecl().ljust(8)
-
-    @classmethod
-    def from_grdecl(cls, order_string):
-        if match_keyword(order_string, "INC"):
-            return cls.INCREASING
-        if match_keyword(order_string, "DEC"):
-            return cls.DECREASING
-
-    @classmethod
-    def from_bgrdecl(cls, unit_string):
-        if isinstance(unit_string, bytes):
-            return cls.from_grdecl(unit_string.decode("ascii"))
-        return cls.from_grdecl(unit_string)
-
-
-@unique
-class Orientation(Enum):
-    """Either up or down, for the grdecl keywords UP and DOWN."""
-
-    UP = auto()
-    DOWN = auto()
-
-    def to_grdecl(self):
-        return self.name
-
-    def to_bgrdecl(self):
-        return self.to_grdecl().ljust(8)
-
-    @classmethod
-    def from_grdecl(cls, orientation_string):
-        if match_keyword(orientation_string, "UP"):
-            return cls.UP
-        if match_keyword(orientation_string, "DOWN"):
-            return cls.DOWN
-        raise ValueError(f"Unknown orientation string {orientation_string}")
-
-    @classmethod
-    def from_bgrdecl(cls, unit_string):
-        if isinstance(unit_string, bytes):
-            return cls.from_grdecl(unit_string.decode("ascii"))
-        return cls.from_grdecl(unit_string)
-
-
-@unique
-class Handedness(Enum):
-    """The handedness of an orientation.
-
-    Eiter left handed or right handed.  Used for the grdecl keywords LEFT and
-    RIGHT.
-    """
-
-    LEFT = auto()
-    RIGHT = auto()
-
-    def to_grdecl(self):
-        return self.name
-
-    def to_bgrdecl(self):
-        return self.to_grdecl().ljust(8)
-
-    @classmethod
-    def from_grdecl(cls, orientation_string):
-        if match_keyword(orientation_string, "LEFT"):
-            return cls.LEFT
-        if match_keyword(orientation_string, "RIGHT"):
-            return cls.RIGHT
-        raise ValueError(f"Unknown handedness string {orientation_string}")
-
-    @classmethod
-    def from_bgrdecl(cls, unit_string):
-        if isinstance(unit_string, bytes):
-            return cls.from_grdecl(unit_string.decode("ascii"))
-        return cls.from_grdecl(unit_string)
-
-
-@unique
-class CoordinateType(Enum):
-    """The coordinate system type given in the SPECGRID keyword.
-
-    This is given by either T or F in the last value of SPECGRID, meaning
-    either cylindrical or cartesian coordinates respectively.
-    """
-
-    CARTESIAN = auto()
-    CYLINDRICAL = auto()
-
-    def to_grdecl(self):
-        if self == CoordinateType.CARTESIAN:
-            return "F"
-        else:
-            return "T"
-
-    def to_bgrdecl(self):
-        if self == CoordinateType.CARTESIAN:
-            return 0
-        else:
-            return 1
-
-    @classmethod
-    def from_bgrdecl(cls, coord_value):
-        if coord_value == 0:
-            return cls.CARTESIAN
-        else:
-            return cls.CYLINDRICAL
-
-    @classmethod
-    def from_grdecl(cls, coord_string):
-        if match_keyword(coord_string, "F"):
-            return cls.CARTESIAN
-        if match_keyword(coord_string, "T"):
-            return cls.CYLINDRICAL
-        raise ValueError(f"Unknown coordinate type {coord_string}")
-
-
-@dataclass
-class GrdeclKeyword:
-    """A general grdecl keyword.
-
-    Gives a general implementation of to/from grdecl which simply recurses on
-    fields.
-    """
-
-    def to_grdecl(self):
-        """Convert the keyword to list of grdecl keyword values.
-        Returns:
-            list of values of the given keyword. ie. The
-            keyword read from "SPECGRID 1 1 1 F" should return
-            [1,1,1,CoordinateType.CYLINDRICAL]
-        """
-        return [value.to_grdecl() for value in astuple(self)]
-
-    def to_bgrdecl(self):
-        return [value.to_bgrdecl() for value in astuple(self)]
-
-    @classmethod
-    def from_bgrdecl(cls, values):
-        object_types = [f.type for f in fields(cls)]
-        return cls(*[typ.from_bgrdecl(val) for val, typ in zip(values, object_types)])
-
-    @classmethod
-    def from_grdecl(cls, values):
-        """Convert list of grdecl keyword values to a keyword.
-        Args:
-            values(list): list of values given after the keyword in
-                the grdecl file.
-        Returns:
-            A GrdeclKeyword constructed from the given values.
-        """
-        object_types = [f.type for f in fields(cls)]
-        return cls(*[typ.from_grdecl(val) for val, typ in zip(values, object_types)])
-
-
-@dataclass
-class MapAxes(GrdeclKeyword):
-    """The mapaxes keyword gives the local coordinate system of the map.
-
-    The map coordinate system is given by a point on the y line, the origin and
-    a point on the x line. ie. The usual coordinate system is given by "MAPAXES
-    0 1 0 0 1 0 /" where the two first values is a point on the y line, the
-    middle two values is the origin, and the last two values is a point on the
-    x line.
-    """
-
-    y_line: Tuple[float, float] = (0.0, 1.0)
-    origin: Tuple[float, float] = (0.0, 0.0)
-    x_line: Tuple[float, float] = (1.0, 0.0)
-
-    def to_grdecl(self):
-        return list(self.y_line) + list(self.origin) + list(self.x_line)
-
-    def to_bgrdecl(self):
-        return self.to_grdecl()
-
-    @classmethod
-    def from_bgrdecl(cls, values):
-        return cls.from_grdecl(values)
-
-    @classmethod
-    def from_grdecl(cls, values):
-        if len(values) != 6:
-            raise ValueError("MAPAXES must contain 6 values")
-        return cls(
-            (float(values[0]), float(values[1])),
-            (float(values[2]), float(values[3])),
-            (float(values[4]), float(values[5])),
-        )
-
-
-@dataclass
-class GdOrient(GrdeclKeyword):
-    """The GDORIENT keyword gives the orientation of the grid.
-
-    The three first values is either increasing or decreasing
-    depending on whether the corresponding dimension has increasing
-    or decreasing coordinates. Then comes the direction of the z dimension,
-    and finally the handedness of the orientation. Defaults to
-    "GDORIENT INC INC INC DOWN RIGHT /".
-    """
-
-    i_order: Order = Order.INCREASING
-    j_order: Order = Order.INCREASING
-    k_order: Order = Order.INCREASING
-    z_direction: Orientation = Orientation.DOWN
-    handedness: Handedness = Handedness.RIGHT
+from ._ecl_grid import (
+    CoordinateType,
+    EclGrid,
+    GdOrient,
+    GrdeclKeyword,
+    GridRelative,
+    GridUnit,
+    MapAxes,
+)
+from ._grdecl_format import open_grdecl
 
 
 @dataclass
@@ -343,51 +93,6 @@ class SpecGrid(GrdeclKeyword):
         if len(values) == 5:
             return cls(*ivalues, CoordinateType.from_grdecl(values[-1]))
         raise ValueError("SPECGRID should have at most 5 values")
-
-
-@dataclass
-class GridUnit(GrdeclKeyword):
-    """
-    Defines the units used for grid dimensions. The first value
-    is a string describing the units used, defaults to METRES, known
-    accepted other units are FIELD and LAB. The last value describes
-    whether the measurements are relative to the map or to the
-    origin of MAPAXES.
-    """
-
-    unit: str = "METRES"
-    grid_relative: GridRelative = GridRelative.ORIGIN
-
-    def to_grdecl(self):
-        return [
-            self.unit,
-            self.grid_relative.to_grdecl(),
-        ]
-
-    def to_bgrdecl(self):
-        return [
-            self.unit.ljust(8),
-            self.grid_relative.to_bgrdecl(),
-        ]
-
-    @classmethod
-    def from_bgrdecl(cls, values):
-        if isinstance(values[0], bytes):
-            return cls.from_grdecl([v.decode("ascii") for v in values])
-        return cls.from_grdecl(values)
-
-    @classmethod
-    def from_grdecl(cls, values):
-        if len(values) == 1:
-            return cls(values[0].rstrip())
-        if len(values) == 2:
-            return cls(
-                values[0].rstrip(),
-                GridRelative.MAP
-                if match_keyword(values[1], "MAP")
-                else GridRelative.ORIGIN,
-            )
-        raise ValueError("GridUnit record must contain either 1 or 2 values")
 
 
 @dataclass
